requirements.txt
colour-science==0.3.16
cycler==0.10.0
decorator==4.4.2
ExifRead==2.3.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==3.4.8.29
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
rawpy==0.16.0
scikit-image==0.18.1
scipy==1.6.2
six==1.15.0
tifffile==2021.4.8
tqdm==4.60.0
pandas==1.2.4
boto3==1.17.54
ipykernel==5.5.3
ipython==7.23.1